fix: handled more refactors
