+++ src/components/LHNOptionsList/OptionRowLHN.js
@@ -6,7 +6,7 @@ import {
     View,
     StyleSheet,
 } from 'react-native';
-import * as localStyles from '../../styles/getOptionRowStyles';
+import * as optionRowStyles from '../../styles/optionRowStyles';
 import styles from '../../styles/styles';
 import * as StyleUtils from '../../styles/StyleUtils';
 import Icon from '../Icon';
@@ -73,10 +73,10 @@ const OptionRowLHN = (props) => {
         ? [styles.ml1, StyleUtils.getBackgroundColorWithOpacityStyle(themeColors.icon, 0.5)]
         : [styles.ml1];
     const alternateTextStyle = StyleUtils.combineStyles(props.viewMode === CONST.OPTION_MODE.COMPACT
-        ? [textStyle, localStyles.compactAlternateTextStyle, styles.textLabelSupporting, styles.optionAlternateTextCompact, styles.ml2]
+        ? [textStyle, optionRowStyles.compactAlternateTextStyle, styles.textLabelSupporting, styles.optionAlternateTextCompact, styles.ml2]
         : [textStyle, styles.optionAlternateText, styles.textLabelSupporting], props.style);
     const contentContainerStyles = props.viewMode === CONST.OPTION_MODE.COMPACT
-        ? [styles.flex1, styles.flexRow, styles.overflowHidden, localStyles.compactContentContainerStyles]
+        ? [styles.flex1, styles.flexRow, styles.overflowHidden, optionRowStyles.compactContentContainerStyles]
         : [styles.flex1];
     const sidebarInnerRowStyle = StyleSheet.flatten(props.viewMode === CONST.OPTION_MODE.COMPACT ? [
         styles.chatLinkRowPressable,
